Change error handler code to return exception
and throw it in _response_for.

This started out as fixing a pylint warning about inconsistent return
statements, as it thought there was a reachable implicit return.
However, this change should also make the stack traces more readable
as there is less noise that is internal to this module. It also makes
the flow of _response_for a bit clearer.